разделение компонентов
