+++ project/src/pages/movie-page-screen/movie-page-screen.tsx
@@ -1,16 +1,16 @@
 import { useEffect } from 'react';
-import { Link, useNavigate, useParams } from 'react-router-dom';
-import { AppRoute, AuthorizationStatus } from '../../const';
+import { useParams } from 'react-router-dom';
 import {Film} from '../../types/film';
 import Header from '../../components/header/header';
 import Footer from '../../components/footer/footer';
 import FilmTabs from '../../components/film-tabs/film-tabs';
 import { useAppDispatch, useAppSelector } from '../../hooks';
-import { fetchFilmByIdAction, fetchCommentsByIdAction, fetchSimilarByIdAction, sendFavoriteStatusAction } from '../../store/api-actions';
-import { authorizationStatusSelector, reviewsSelector, movieSelector, similarFilmsSelector } from '../../store/selectors';
+import { fetchFilmByIdAction, fetchCommentsByIdAction, fetchSimilarByIdAction } from '../../store/api-actions';
+import { reviewsSelector, movieSelector, similarFilmsSelector } from '../../store/selectors';
 import LoadingScreen from '../loading-screen/loading-screen';
 import DefaultLoader from '../../components/loader/loader';
 import SimilarFilms from '../../components/similar-films/similar-films';
+import MovieInfo from '../../components/movie-info/movie-info';
 
 
 type MoviePageProps = {
@@ -20,12 +20,10 @@ type MoviePageProps = {
 function MoviePage(props: MoviePageProps): JSX.Element {
 
   const params = useParams();
-  const navigate = useNavigate();
   const dispatch = useAppDispatch();
   const movieInfo = useAppSelector(movieSelector);
   const filmReviews = useAppSelector(reviewsSelector);
   const similarFilms = useAppSelector(similarFilmsSelector);
-  const authorizationStatus = useAppSelector(authorizationStatusSelector);
 
   useEffect(() => {
     if (!movieInfo.isError) {
@@ -37,13 +35,6 @@ function MoviePage(props: MoviePageProps): JSX.Element {
 
   const isFavorite = props.favoriteFilms.map((film) => film.id).includes(Number(params.id));
 
-  const buttonClickHandler = () => {
-    dispatch(sendFavoriteStatusAction({
-      status: Number(!isFavorite),
-      filmId: String(movieInfo.data?.id)
-    }));
-  };
-
   if (movieInfo.data === null) {
     return <LoadingScreen/>;
   }
@@ -51,42 +42,10 @@ function MoviePage(props: MoviePageProps): JSX.Element {
   return (
     <>
       <section className="film-card film-card--full">
-        <div className="film-card__hero">
-          <div className="film-card__bg">
-            <img src={movieInfo.data.backgroundImage} alt={movieInfo.data.name} />
-          </div>
 
-          <h1 className="visually-hidden">WTW</h1>
+        <MovieInfo favoriteFilms={props.favoriteFilms} movieInfo={movieInfo.data} isFavorite={isFavorite}>
           <Header/>
-          <div className="film-card__wrap">
-            <div className="film-card__desc">
-              <h2 className="film-card__title">{movieInfo.data.name}</h2>
-              <p className="film-card__meta">
-                <span className="film-card__genre">{movieInfo.data.genre}</span>
-                <span className="film-card__year">{movieInfo.data.released}</span>
-              </p>
-
-              <div className="film-card__buttons">
-                <button className="btn btn--play film-card__button" type="button" onClick={() => navigate(`/player/${String(movieInfo.data?.id)}`)}>
-                  <svg viewBox="0 0 19 19" width="19" height="19">
-                    <use xlinkHref="#play-s"></use>
-                  </svg>
-                  <span>Play</span>
-                </button>
-                <button className="btn btn--list film-card__button" type="button" onClick={buttonClickHandler}>
-                  <svg viewBox={isFavorite ? '0 0 19 19' : '0 0 19 20'} width="19" height={isFavorite ? '19' : '20'}>
-                    <use xlinkHref={isFavorite ? '#in-list' : '#add'}></use>
-                  </svg>
-                  <span>My list</span>
-                  <span className="film-card__count">{props.favoriteFilms.length}</span>
-                </button>
-                {authorizationStatus === AuthorizationStatus.Auth ?
-                  <Link to={`${AppRoute.AddReview}`} className="btn film-card__button">Add review</Link>
-                  : null}
-              </div>
-            </div>
-          </div>
-        </div>
+        </MovieInfo>
 
         <div className="film-card__wrap film-card__translate-top">
           <div className="film-card__info">
